#171 Combine closely connected components
